Create test cases for all controllers #28 Added test for PromotionController. Modified PromotionController to fulfil coding standards.

+++ src/controllers/PromotionController.php
@@ -9,8 +9,8 @@ use Illuminate\Support\Facades\View;
 use Cartalyst\Sentry\Facades\Laravel\Sentry;
 use DateTime;
 
-class PromotionController extends BaseController {
-
+class PromotionController extends BaseController
+{
     protected $model;
     
     public function __construct(Promotion $promotion)
@@ -30,27 +30,28 @@ class PromotionController extends BaseController {
         return View::make('redminportal::promotions/create');
     }
     
-    public function getEdit($id)
+    public function getEdit($sid)
     {
         // Find the promotion using the user id
-        $promotion = Promotion::find($id);
+        $promotion = Promotion::find($sid);
         
-		if(empty($promotion->options))
-        {
+        if ($promotion == null) {
+            return \View::make('redminportal::pages/404');
+        }
+        
+        if (empty($promotion->options)) {
             $promotion_cn = (object) array(
                 'name'                  => $promotion->name,
                 'short_description'     => $promotion->short_description,
                 'long_description'      => $promotion->long_description
             );
-        }
-        else
-        {
+        } else {
             $promotion_cn = json_decode($promotion->options);
         }
-		
+        
         return View::make('redminportal::promotions/edit')
             ->with('promotion', $promotion)
-			->with('promotion_cn', $promotion_cn)
+            ->with('promotion_cn', $promotion_cn)
             ->with('imageUrl', 'assets/img/promotions/')
             ->with('start_date', new DateTime($promotion->start_date))
             ->with('end_date', new DateTime($promotion->end_date));
@@ -58,11 +59,8 @@ class PromotionController extends BaseController {
     
     public function postStore()
     {
-        $id = Input::get('id');
+        $sid = Input::get('id');
         
-        /*
-         * Validate
-         */
         $rules = array(
             'image'             => 'mimes:jpg,jpeg,png,gif|max:500',
             'name'              => 'required',
@@ -74,28 +72,37 @@ class PromotionController extends BaseController {
         
         $validation = Validator::make(Input::all(), $rules);
          
-        if( $validation->passes() )
-        {
+        if ($validation->passes()) {
             $name               = Input::get('name');
             $short_description  = Input::get('short_description');
             $long_description   = Input::get('long_description');
             $image              = Input::file('image');
-            $active             = (Input::get('active') == '' ? FALSE : TRUE);
+            $active             = (Input::get('active') == '' ? false : true);
             
-			$cn_name               = Input::get('cn_name');
+            $cn_name               = Input::get('cn_name');
             $cn_short_description  = Input::get('cn_short_description');
             $cn_long_description   = Input::get('cn_long_description');
-			
-			$options = array(
+            
+            $options = array(
                 'name'                  => $cn_name,
                 'short_description'     => $cn_short_description,
                 'long_description'      => $cn_long_description
             );
-			
+            
             $start_date = DateTime::createFromFormat('d/m/Y', Input::get('start_date'));
             $end_date   = DateTime::createFromFormat('d/m/Y', Input::get('end_date'));
             
-            $promotion = (isset($id) ? Promotion::find($id) : new Promotion);
+            $promotion = (isset($sid) ? Promotion::find($sid) : new Promotion);
+            
+            if ($promotion == null) {
+                $errors = new \Illuminate\Support\MessageBag;
+                $errors->add(
+                    'editError',
+                    "The promotion cannot be found because it does not exist or may have been deleted."
+                );
+                return \Redirect::to('/admin/promotions')->withErrors($errors);
+            }
+            
             $promotion->name = $name;
             $promotion->start_date = $start_date;
             $promotion->end_date = $end_date;
@@ -103,13 +110,14 @@ class PromotionController extends BaseController {
             $promotion->long_description = $long_description;
             $promotion->active = $active;
             $promotion->options = json_encode($options);
-			
+            
             $promotion->save();
             
-            if(Input::hasFile('image'))
-            {
+            if (Input::hasFile('image')) {
                 // Delete all existing images for edit
-                if(isset($id)) $promotion->deleteAllImages();
+                if (isset($sid)) {
+                    $promotion->deleteAllImages();
+                }
                 
                 //set the name of the file
                 $originalFilename = $image->getClientOriginalName();
@@ -118,8 +126,7 @@ class PromotionController extends BaseController {
                 //Upload the file
                 $isSuccess = $image->move('assets/img/promotions', $filename);
                 
-                if( $isSuccess )
-                {
+                if ($isSuccess) {
                     // create photo
                     $newimage = new Image;
                     $newimage->path = $filename;
@@ -128,15 +135,11 @@ class PromotionController extends BaseController {
                     $promotion->images()->save($newimage);
                 }
             }
-
-        }//if it validate
-        else {
-            if(isset($id))
-            {
-                return Redirect::to('admin/promotions/edit/' . $id)->withErrors($validation)->withInput();
-            }
-            else
-            {
+        //if it validate
+        } else {
+            if (isset($sid)) {
+                return Redirect::to('admin/promotions/edit/' . $sid)->withErrors($validation)->withInput();
+            } else {
                 return Redirect::to('admin/promotions/create')->withErrors($validation)->withInput();
             }
         }
@@ -144,10 +147,16 @@ class PromotionController extends BaseController {
         return Redirect::to('admin/promotions');
     }
     
-    public function getDelete($id)
+    public function getDelete($sid)
     {
         // Find the promotion using the id
-        $promotion = Promotion::find($id);
+        $promotion = Promotion::find($sid);
+        
+        if ($promotion == null) {
+            $errors = new \Illuminate\Support\MessageBag;
+            $errors->add('deleteError', "We are having problem deleting this entry. Please try again.");
+            return \Redirect::to('admin/promotions')->withErrors($errors);
+        }
         
         // Delete all images
         $promotion->deleteAllImages();
@@ -157,5 +166,4 @@ class PromotionController extends BaseController {
 
         return Redirect::to('admin/promotions');
     }
-
 }
